HPCC-9799 Status
Working BUT just about to deviate between bottom up construction of nodes in
syntax tree and top down.

e.g. setRoot (ast) to a pointer to a node created in some later production or
create the node in setRoot and then then just fill pointers for left and right.
Bonus of this is that one can than use $$->bifurcate syntax rather than
$$ = parser->bifurcate.
Signed-off-by: jamienoss <james.noss@lexisnexis.com>